MAINT: Moving erros.py into helper direcotry
Moving this module into the helper directory better explains what it's for and that it is a shared module
