+++ cloud-chatops/tests/test_dev.py
@@ -5,7 +5,7 @@ import pytest
 
 from helper.data import User
 from dev import run_methods, call_method, parse_args
-from errors import NoTestCase
+from helper.errors import NoTestCase
 
 
 MOCK_USER = User(
